// src/shared/generated/CoreEffectResult.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CoreErrorPayload } from "./CoreErrorPayload";

export type CoreEffectResult = { effectId: string, operationId: string, ok: boolean, valueJson: string | null, error: CoreErrorPayload | null, };